requirements.txt
biopython==1.81
freesasa
stable-baselines3
peptidebuilder
openmm